♻️ Refactor phase management to use shared table component
Replaces duplicated phase management table logic in innovation-package and reporting modules with a new shared PhaseManagementTable component. Removes local table markup, related SCSS, and test files from both modules, and updates module imports to use the shared component. This improves maintainability and consistency across phase management features.
