+++ onecgiar-pr-client/src/app/pages/admin-section/pages/phase-management/pages/innovation-package/innovation-package.component.html
@@ -1,136 +1,12 @@
-<br>
-<div class="setting_title">Phases</div>
-
-<div class="local_container phase_table">
-  <p-table sortField="result_code" [sortOrder]="-1" [value]="this.phaseList" scrollHeight="calc(100vh - 303px)"
-    responsiveLayout="scroll" [paginator]="!!(this.phaseList)?.length" [rows]="10"
-    styleClass="p-datatable-gridlines p-datatable-sm" selectionMode="single" [rowsPerPageOptions]="[10,50,100]">
-    <ng-template pTemplate="header">
-      <tr>
-        <th *ngFor="let column of columnOrder" pSortableColumn="{{column.attr}}" id="{{column.attr}}">
-          <p-sortIcon field="{{column.attr}}"></p-sortIcon>{{column.title}}
-        </th>
-        <th id="actions">Actions</th>
-      </tr>
-    </ng-template>
-    <ng-template pTemplate="body" let-phase>
-      <tr>
-        <td [style.width.px]="30">
-          {{phase?.id}}
-        </td>
-        <td [style.width.px]="130">
-          <div class="text" *ngIf="!phase.editing">{{phase?.phase_name}}</div>
-          <app-pr-input *ngIf="phase.editing" placeholder="Provide input here..." type="text"
-            [(ngModel)]="phase.phase_name_ts">
-          </app-pr-input>
-        </td>
-        <td [style.width.px]="110">
-          <div class="text" *ngIf="!phase.isNew">{{phase?.phase_year}}</div>
-          <app-pr-select *ngIf="phase.isNew" class="segment_title_margin" [options]="this.resultYearsList"
-            [disableOptions]="disablePreviousYear()" optionLabel="year" optionValue="year" placeholder="Select year"
-            [(ngModel)]="phase.phase_year_ts">
-          </app-pr-select>
-        </td>
-        <td [style.width.px]="180">
-          <div class="text" *ngIf="!phase.editing">
-            {{phase?.obj_portfolio?.name || ''}}
-          </div>
-          <app-pr-select *ngIf="phase.editing"
-            class="segment_title_margin"
-            [options]="portfolioList"
-            optionLabel="name"
-            optionValue="id"
-            placeholder="Select portfolio"
-            [(ngModel)]="phase.portfolio_id_ts">
-          </app-pr-select>
-        <td>
-          <div class="text" *ngIf="!phase.isNew">{{phase?.obj_reporting_phase?.phase_name}}</div>
-          <app-pr-select *ngIf="phase.isNew" class="segment_title_margin" [options]="this.reportingPhasesList"
-            optionLabel="phase_name" optionValue="id" placeholder="Select phase" [(ngModel)]="phase.reporting_phase_ts">
-          </app-pr-select>
-        </td>
-        <td [style.width.px]="180">
-
-          <div class="text" *ngIf="!phase.isNew">{{this.getTocPhaseName(phase?.toc_pahse_id)}}</div>
-          <app-pr-select *ngIf="phase.isNew" class="segment_title_margin" [options]="this.tocPhaseList"
-            optionLabel="fullText" optionValue="phase_id" placeholder="Select phase"
-            [(ngModel)]="phase.toc_pahse_id_ts">
-          </app-pr-select>
-        </td>
-        <td [style.width.px]="120">
-
-          <div class="text" *ngIf="!phase.isNew"> {{phase?.start_date | date: 'YYYY-MM-dd'}}</div>
-          <p-calendar [touchUI]="false" placeholder="Start date" *ngIf="phase.isNew" [(ngModel)]="phase.start_date_ts"
-            dateFormat="yy-mm-dd"></p-calendar>
-        </td>
-        <td [style.width.px]="120">
-
-          <div class="text" *ngIf="!phase.isNew">{{phase?.end_date| date: 'YYYY-MM-dd' }}</div>
-          <p-calendar [touchUI]="false" placeholder="End date" *ngIf="phase.isNew" [(ngModel)]="phase.end_date_ts"
-            dateFormat="yy-mm-dd"></p-calendar>
-        </td>
-        <td [style.width.px]="90">
-          <div class="text" *ngIf="!phase.editing">
-            <div class="tag_container" [ngClass]="phase?.status ? 'open':'closed'">
-              {{phase?.status ? 'Open':'Closed'}}
-            </div>
-          </div>
-          <div class="text" *ngIf="phase.isNew">
-            <div class="tag_container" [ngClass]="'closed'">
-              Closed
-            </div>
-          </div>
-          <app-pr-select *ngIf="phase.editing  && !phase.isNew" class="segment_title_margin" [options]="this.status"
-            optionLabel="name" optionValue="status" placeholder="Select status" [(ngModel)]="phase.status_ts">
-          </app-pr-select>
-        </td>
-        <td>
-          <div class="text" *ngIf="!phase.editing">{{phase?.obj_previous_phase? phase?.obj_previous_phase?.phase_name:
-            'N/A'}}</div>
-          <app-pr-select *ngIf="phase.editing" class="segment_title_margin" [options]="this.previousPhaseList"
-            [disableOptions]="onlyPreviousPhase(phase)" optionLabel="phase_name" optionValue="id"
-            placeholder="Select previous phase" [(ngModel)]="phase.previous_phase_ts">
-          </app-pr-select>
-        </td>
-
-        <td [style.width.px]="85">
-          <div class="actions" *ngIf="!phase.editing">
-            <div class="action_button edit" (click)="getFeedback() ? null:phase.editing = true"
-              [ngClass]="{'disabled': getFeedback()}"><i class="material-icons-round"
-                [pTooltip]="getFeedback()?this.disabledActionsText:null" tooltipPosition="top">edit</i>
-            </div>
-            <div class="action_button delete" [ngClass]="{'disabled': getFeedback() ||!phase?.can_be_deleted}"
-              (click)="getFeedback()?null:(phase?.can_be_deleted ?this.deletePhase(phase):null)"
-              [pTooltip]="getFeedback()?this.disabledActionsText:(!phase?.can_be_deleted?'The phase has active results therefore cannot be eliminated':'')"
-              tooltipPosition="top"><i class="material-icons-round">delete</i>
-            </div>
-          </div>
-          <div class="actions" *ngIf="phase.editing">
-            <div class="action_button save" tooltipPosition="top"
-              [ngClass]="{'disabled': phase?.isNew  &&  getMandatoryIncompleteFields(phase)}"
-              [pTooltip]="phase?.isNew ? getMandatoryIncompleteFields(phase):null" [escape]="false"
-              (click)="phase?.isNew  &&  getMandatoryIncompleteFields(phase) ?null:(phase?.isNew ? createPhase(phase) :savePhase(phase))">
-              <i class="material-icons-round">{{phase?.isNew ? 'add_circle' : 'save'}}</i>
-            </div>
-            <div class="action_button cancel" (click)="cancelAction(phase); this.updateVariablesToSave(phase)">
-              <i class="material-icons-round">close</i>
-            </div>
-          </div>
-        </td>
-      </tr>
-    </ng-template>
-  </p-table>
-</div>
-
-
-<div class="center_flex_100">
-  <div class="add_new_phase_button" [ngClass]="{'disabled': getFeedback()}">
-    <div class="feedback" *ngIf="getFeedback()">{{getFeedback()}}</div>
-    <div class="button" (click)="getFeedback()?'':addNewPhase()">Add new phase</div>
+<app-phase-management-table
+  [moduleType]="moduleType"
+  [appModuleId]="appModuleId"
+  [showReportingPhaseColumn]="true"
+  [placeholderText]="'Provide input here...'"
+  (onPhaseUpdate)="onPhaseUpdate()">
+  
+  <div style="margin-top: 2rem;" *ngIf="this.api.dataControlSE.previousIPSRPhase.phaseName">
+    <div class="setting_title">Other functionalities</div>
+    <app-other-functionalities [replicateIPSR]="true"></app-other-functionalities>
   </div>
-</div>
-
-<div style="margin-top: 2rem;" *ngIf="this.api.dataControlSE.previousIPSRPhase.phaseName">
-  <div class="setting_title">Other functionalities</div>
-  <app-other-functionalities [replicateIPSR]="true"></app-other-functionalities>
-</div>
+</app-phase-management-table>
