+++ onecgiar-pr-client/src/app/pages/admin-section/pages/phase-management/pages/reporting/reporting.component.scss
@@ -1,135 +1,2 @@
-.setting_title {
-  font-size: 1.5rem;
-  font-weight: 500;
-  margin-bottom: 10px;
-  padding-left: 30px;
-}
-
-.local_container {
-  padding: 30px;
-}
-
-.actions {
-  min-width: 85px;
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  justify-content: center;
-
-  .action_button {
-    $edit-color: #37db68;
-    $cancel-color: #f75252;
-    $delete-color: #f75252;
-    $save-color: #5286f7;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 5px;
-    border: 1.5px solid;
-    width: max-content;
-    cursor: pointer;
-
-    i {
-      padding: 3px;
-      font-size: 1.6rem;
-    }
-
-    .text {
-      padding-left: 5px;
-      margin-bottom: 2px;
-      font-weight: 500;
-    }
-
-    &.edit {
-      border-color: $edit-color;
-
-      i {
-        color: $edit-color;
-      }
-    }
-
-    &.delete {
-      border-color: $delete-color;
-
-      i {
-        color: $delete-color;
-      }
-    }
-
-    &.save {
-      border-color: $save-color;
-
-      i,
-      .text {
-        color: $save-color;
-      }
-    }
-
-    &.cancel {
-      border-color: $cancel-color;
-
-      i,
-      .text {
-        color: $cancel-color;
-      }
-    }
-
-    $disabled-color: #b8b8b8;
-
-    &.disabled {
-      border-color: $disabled-color !important;
-      cursor: not-allowed;
-
-      i,
-      .text {
-        color: $disabled-color !important;
-      }
-    }
-  }
-}
-
-.tag_container {
-  color: white;
-  border-radius: 5px;
-  padding: 5px 10px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  &.open {
-    background-color: #23be4a;
-  }
-
-  &.closed {
-    background-color: #d13838;
-  }
-}
-
-.add_new_phase_button {
-  width: auto;
-  margin: 0px 100px;
-
-  .feedback {
-    background-color: rgb(248, 79, 79);
-    color: white;
-    padding: 3px 10px;
-  }
-
-  .button {
-    background-color: var(--pr-color-primary);
-    color: white;
-    padding: 10px 20px;
-    border-radius: 0px 0px 6px 6px;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-}
-
-.disabled {
-  .button {
-    background-color: #92949c !important;
-    cursor: not-allowed !important;
-  }
-}
+// Los estilos ahora están en el componente compartido
+// Este archivo puede quedar vacío o con estilos específicos de reporting si los hay
