--- onecgiar-pr-client/src/app/pages/admin-section/pages/phase-management/pages/reporting/reporting.component.spec.ts
@@ -1,551 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { of, throwError } from 'rxjs';
-import { ReportingComponent } from './reporting.component';
-import { ResultsApiService } from '../../../../../../shared/services/api/results-api.service';
-import { CustomizedAlertsFeService } from '../../../../../../shared/services/customized-alerts-fe.service';
-import { PhasesService } from '../../../../../../shared/services/global/phases.service';
-import { ApiService } from '../../../../../../shared/services/api/api.service';
-import { Phase } from '../../../../../../shared/interfaces/phase.interface';
-import { StatusPhaseEnum, ModuleTypeEnum } from '../../../../../../shared/enum/api.enum';
-
-describe('ReportingComponent', () => {
-  let component: ReportingComponent;
-  let fixture: ComponentFixture<ReportingComponent>;
-  let mockResultsApiService: any;
-  let mockCustomizedAlertsFeService: any;
-  let mockPhasesService: any;
-  let mockApiService: any;
-
-  const mockPhaseData = [
-    {
-      id: 1,
-      phase_name: 'Test Phase 1',
-      phase_year: 2024,
-      toc_pahse_id: 1,
-      start_date: '2024-01-01',
-      end_date: '2024-12-31',
-      status: true,
-      previous_phase: null,
-      can_be_deleted: true,
-      obj_previous_phase: null
-    },
-    {
-      id: 2,
-      phase_name: 'Test Phase 2',
-      phase_year: 2023,
-      toc_pahse_id: 2,
-      start_date: '2023-01-01',
-      end_date: '2023-12-31',
-      status: false,
-      previous_phase: 1,
-      can_be_deleted: false,
-      obj_previous_phase: { phase_name: 'Test Phase 1' }
-    }
-  ];
-
-  const mockTocPhases = [
-    { phase_id: 1, name: 'TOC Phase 1', status: 'Active' },
-    { phase_id: 2, name: 'TOC Phase 2', status: 'Inactive' }
-  ];
-
-  const mockPortfolios = [
-    { id: 2, name: 'CGIAR portfolio 2022-2024​', startDate: 2022, endDate: 2024, isActive: 1 },
-    { id: 3, name: 'CGIAR portfolio 2025-2030', startDate: 2025, endDate: 2030, isActive: 0 }
-  ];
-
-  const mockResultYears = [{ year: 2024 }, { year: 2023 }, { year: 2022 }];
-
-  beforeEach(async () => {
-    mockResultsApiService = {
-      GET_versioning: jest.fn().mockReturnValue(of({ response: mockPhaseData })),
-      GET_tocPhases: jest.fn().mockReturnValue(of({ response: mockTocPhases })),
-      GET_resultYears: jest.fn().mockReturnValue(of({ response: mockResultYears })),
-      PATCH_updatePhase: jest.fn().mockReturnValue(of({})),
-      POST_createPhase: jest.fn().mockReturnValue(of({})),
-      DELETE_updatePhase: jest.fn().mockReturnValue(of({}))
-    };
-
-    mockCustomizedAlertsFeService = {
-      show: jest.fn()
-    };
-
-    mockPhasesService = {
-      getNewPhases: jest.fn()
-    };
-
-    mockApiService = {
-      dataControlSE: {
-        getCurrentPhases: jest.fn(),
-        previousReportingPhase: { phaseName: 'Previous Phase' }
-      }
-    };
-
-    mockResultsApiService.GET_portfolioList = jest.fn().mockReturnValue(of(mockPortfolios));
-
-    await TestBed.configureTestingModule({
-      declarations: [ReportingComponent],
-      providers: [
-        { provide: ResultsApiService, useValue: mockResultsApiService },
-        { provide: CustomizedAlertsFeService, useValue: mockCustomizedAlertsFeService },
-        { provide: PhasesService, useValue: mockPhasesService },
-        { provide: ApiService, useValue: mockApiService }
-      ]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(ReportingComponent);
-    component = fixture.componentInstance;
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-
-  it('should initialize component with default values', () => {
-    expect(component.phaseList).toEqual([]);
-    expect(component.previousPhaseList).toEqual([]);
-    expect(component.tocPhaseList).toEqual([]);
-    expect(component.resultYearsList).toEqual([]);
-    expect(component.textToFind).toBe('');
-    expect(component.disabledActionsText).toBe('Finish editing the phase to be able to edit or delete this phase.');
-  });
-
-  it('should have correct column order configuration', () => {
-    expect(component.columnOrder).toEqual([
-      { title: '#', attr: 'id' },
-      { title: 'Name', attr: 'phase_name' },
-      { title: 'Reporting year', attr: 'phase_year' },
-      { title: 'Portfolio', attr: 'portfolio_id' },
-      { title: 'Toc phase', attr: 'toc_pahse_id' },
-      { title: 'Start date', attr: 'start_date' },
-      { title: 'End date', attr: 'end_date' },
-      { title: 'Status', attr: 'status' },
-      { title: 'Previous phase', attr: 'obj_previous_phase' }
-    ]);
-  });
-
-  it('should have correct status options', () => {
-    expect(component.status).toEqual([
-      { status: true, name: 'Open' },
-      { status: false, name: 'Closed' }
-    ]);
-  });
-
-  describe('ngOnInit', () => {
-    it('should call all initialization methods', () => {
-      jest.spyOn(component, 'getAllPhases');
-      jest.spyOn(component, 'getTocPhases');
-      jest.spyOn(component, 'get_resultYears');
-
-      component.ngOnInit();
-
-      expect(component.getAllPhases).toHaveBeenCalled();
-      expect(component.getTocPhases).toHaveBeenCalled();
-      expect(component.get_resultYears).toHaveBeenCalled();
-      expect(mockApiService.dataControlSE.getCurrentPhases).toHaveBeenCalled();
-    });
-  });
-
-  describe('getAllPhases', () => {
-    it('should fetch and process phases data', () => {
-      component.getAllPhases();
-
-      expect(mockResultsApiService.GET_versioning).toHaveBeenCalledWith(StatusPhaseEnum.ALL, ModuleTypeEnum.REPORTING);
-      expect(component.phaseList.length).toBe(2);
-      expect(component.previousPhaseList.length).toBe(3); // 2 phases + N/A option
-      expect(component.previousPhaseList[2]).toEqual({ phase_name: 'N/A', id: null });
-    });
-
-    it('should update variables to save for each phase', () => {
-      jest.spyOn(component, 'updateVariablesToSave');
-      component.getAllPhases();
-
-      expect(component.updateVariablesToSave).toHaveBeenCalledTimes(2);
-    });
-  });
-
-  describe('getTocPhases', () => {
-    it('should fetch and process TOC phases', () => {
-      component.getTocPhases();
-
-      expect(mockResultsApiService.GET_tocPhases).toHaveBeenCalled();
-      expect(component.tocPhaseList.length).toBe(2);
-      expect(component.tocPhaseList[0]).toEqual({
-        phase_id: 1,
-        name: 'TOC Phase 1',
-        status: 'Active',
-        fullText: 'TOC Phase 1 - Active'
-      });
-    });
-  });
-
-  describe('get_resultYears', () => {
-    it('should fetch result years', () => {
-      component.get_resultYears();
-
-      expect(mockResultsApiService.GET_resultYears).toHaveBeenCalled();
-      expect(component.resultYearsList).toEqual(mockResultYears);
-    });
-  });
-
-  describe('updateVariablesToSave', () => {
-    it('should copy main variables to temporary save variables', () => {
-      const phase: any = {
-        phase_name: 'Test',
-        phase_year: 2024,
-        toc_pahse_id: 1,
-        start_date: '2024-01-01',
-        end_date: '2024-12-31',
-        status: true,
-        previous_phase: null
-      };
-
-      component.updateVariablesToSave(phase);
-
-      expect(phase.phase_name_ts).toBe('Test');
-      expect(phase.phase_year_ts).toBe(2024);
-      expect(phase.toc_pahse_id_ts).toBe(1);
-      expect(phase.start_date_ts).toBe('2024-01-01');
-      expect(phase.end_date_ts).toBe('2024-12-31');
-      expect(phase.status_ts).toBe(true);
-      expect(phase.previous_phase_ts).toBe(null);
-    });
-  });
-
-  describe('updateMainVariables', () => {
-    it('should copy temporary save variables back to main variables', () => {
-      const phase: any = {
-        phase_name_ts: 'Test Updated',
-        phase_year_ts: 2025,
-        toc_pahse_id_ts: 2,
-        start_date_ts: '2025-01-01',
-        end_date_ts: '2025-12-31',
-        status_ts: false,
-        previous_phase_ts: 1
-      };
-
-      component.updateMainVariables(phase);
-
-      expect(phase.phase_name).toBe('Test Updated');
-      expect(phase.phase_year).toBe(2025);
-      expect(phase.toc_pahse_id).toBe(2);
-      expect(phase.start_date).toBe('2025-01-01');
-      expect(phase.end_date).toBe('2025-12-31');
-      expect(phase.status).toBe(false);
-      expect(phase.previous_phase).toBe(1);
-    });
-  });
-
-  describe('getMandatoryIncompleteFields', () => {
-    it('should return empty string when all required fields are present', () => {
-      const phase = {
-        phase_name_ts: 'Test',
-        phase_year_ts: 2024,
-        toc_pahse_id_ts: 1,
-        start_date_ts: '2024-01-01',
-        end_date_ts: '2024-12-31',
-        portfolio_id_ts: 2
-      };
-
-      const result = component.getMandatoryIncompleteFields(phase);
-      expect(result).toBe('');
-    });
-
-    it('should return error message for missing required fields', () => {
-      const phase = {
-        phase_name_ts: '',
-        phase_year_ts: null,
-        toc_pahse_id_ts: null,
-        start_date_ts: '',
-        end_date_ts: '',
-        portfolio_id_ts: null
-      };
-
-      const result = component.getMandatoryIncompleteFields(phase);
-      expect(result).toContain('<strong> Name </strong> is required to create');
-      expect(result).toContain('<strong> Reporting year </strong> is required to create');
-      expect(result).toContain('<strong> Toc phase </strong> is required to create');
-      expect(result).toContain('<strong> Start date </strong> is required to create');
-      expect(result).toContain('<strong> End date </strong>is required to create');
-      expect(result).toContain('<strong> Portfolio </strong> is required to create');
-
-    });
-  });
-
-  describe('savePhase', () => {
-    it('should successfully save an existing phase', () => {
-      const phase = { id: 1, editing: true };
-      jest.spyOn(component, 'updateMainVariables');
-      jest.spyOn(component, 'getAllPhases');
-
-      component.savePhase(phase);
-
-      expect(component.updateMainVariables).toHaveBeenCalledWith(phase);
-      expect(mockResultsApiService.PATCH_updatePhase).toHaveBeenCalledWith(1, phase);
-      expect(component.getAllPhases).toHaveBeenCalled();
-      expect(mockCustomizedAlertsFeService.show).toHaveBeenCalledWith({
-        id: 'manage-phase-save',
-        title: 'Phase saved',
-        status: 'success',
-        closeIn: 500
-      });
-      expect(mockPhasesService.getNewPhases).toHaveBeenCalled();
-      expect(mockApiService.dataControlSE.getCurrentPhases).toHaveBeenCalled();
-      expect(phase.editing).toBe(false);
-    });
-
-    it('should handle save phase error', () => {
-      const phase = { id: 1, editing: true };
-      const error = { error: { message: 'Save failed' } };
-      mockResultsApiService.PATCH_updatePhase.mockReturnValue(throwError(error));
-      jest.spyOn(console, 'error').mockImplementation();
-
-      component.savePhase(phase);
-
-      expect(console.error).toHaveBeenCalledWith(error);
-    });
-  });
-
-  describe('createPhase', () => {
-    it('should successfully create a new phase', () => {
-      const phase: any = { isNew: true };
-      jest.spyOn(component, 'updateMainVariables');
-      jest.spyOn(component, 'getAllPhases');
-
-      component.createPhase(phase);
-
-      expect(phase.app_module_id).toBe(1);
-      expect(component.updateMainVariables).toHaveBeenCalledWith(phase);
-      expect(mockResultsApiService.POST_createPhase).toHaveBeenCalledWith(phase);
-      expect(component.getAllPhases).toHaveBeenCalled();
-      expect(mockCustomizedAlertsFeService.show).toHaveBeenCalledWith({
-        id: 'manage-phase-save',
-        title: 'Phase created',
-        status: 'success',
-        closeIn: 500
-      });
-      expect(phase.isNew).toBe(false);
-      expect(mockPhasesService.getNewPhases).toHaveBeenCalled();
-      expect(mockApiService.dataControlSE.getCurrentPhases).toHaveBeenCalled();
-    });
-
-    it('should handle create phase error', () => {
-      const phase = { isNew: true };
-      const error = { error: { message: 'Create failed' } };
-      mockResultsApiService.POST_createPhase.mockReturnValue(throwError(error));
-      jest.spyOn(console, 'error').mockImplementation();
-
-      component.createPhase(phase);
-
-      expect(console.error).toHaveBeenCalledWith(error);
-      expect(mockCustomizedAlertsFeService.show).toHaveBeenCalledWith({
-        id: 'manage-error',
-        title: 'Create phase',
-        description: 'Create failed',
-        status: 'error',
-        closeIn: 500
-      });
-    });
-  });
-
-  describe('deletePhase', () => {
-    it('should show confirmation dialog and delete phase on confirmation', () => {
-      const phase = { id: 1 };
-      jest.spyOn(component, 'getAllPhases');
-
-      // Mock the confirmation callback
-      mockCustomizedAlertsFeService.show.mockImplementation((config, callback) => {
-        callback(); // Simulate user confirming
-      });
-
-      component.deletePhase(phase);
-
-      expect(mockCustomizedAlertsFeService.show).toHaveBeenCalledWith(
-        {
-          id: 'manage-phase',
-          title: 'Delete phase',
-          description: 'Are you sure you want to delete the current phase?',
-          status: 'warning',
-          confirmText: 'Yes, delete'
-        },
-        expect.any(Function)
-      );
-
-      expect(mockResultsApiService.DELETE_updatePhase).toHaveBeenCalledWith(1);
-      expect(component.getAllPhases).toHaveBeenCalled();
-    });
-
-    it('should handle delete phase error', () => {
-      const phase = { id: 1 };
-      const error = { error: { message: 'Delete failed' } };
-      mockResultsApiService.DELETE_updatePhase.mockReturnValue(throwError(error));
-      jest.spyOn(console, 'error').mockImplementation();
-
-      mockCustomizedAlertsFeService.show.mockImplementation((config, callback) => {
-        callback();
-      });
-
-      component.deletePhase(phase);
-
-      expect(console.error).toHaveBeenCalledWith(error);
-      expect(mockCustomizedAlertsFeService.show).toHaveBeenCalledWith({
-        id: 'manage-error',
-        title: 'Delete phase',
-        description: 'Delete failed',
-        status: 'error',
-        closeIn: 500
-      });
-    });
-  });
-
-  describe('getTocPhaseName', () => {
-    it('should return formatted TOC phase name', () => {
-      component.tocPhaseList = mockTocPhases.map(phase => ({
-        ...phase,
-        fullText: `${phase.name} - ${phase.status}`
-      }));
-
-      const result = component.getTocPhaseName(1);
-      expect(result).toBe('TOC Phase 1 - Active');
-    });
-
-    it('should handle non-existent TOC phase ID', () => {
-      component.tocPhaseList = mockTocPhases;
-
-      const result = component.getTocPhaseName(999);
-      expect(result).toBe('undefined - undefined');
-    });
-  });
-
-  describe('getFeedback', () => {
-    it('should return message when there is a new phase', () => {
-      component.phaseList = [{ isNew: true }];
-
-      const result = component.getFeedback();
-      expect(result).toBe('Create or cancel to add a new phase');
-    });
-
-    it('should return message when there is an editing phase', () => {
-      component.phaseList = [{ editing: true }];
-
-      const result = component.getFeedback();
-      expect(result).toBe('Save or cancel to add a new phase');
-    });
-
-    it('should return empty string when no phases are being edited or created', () => {
-      component.phaseList = [{ isNew: false, editing: false }];
-
-      const result = component.getFeedback();
-      expect(result).toBe('');
-    });
-  });
-
-  describe('addNewPhase', () => {
-    it('should add a new phase to the phase list', () => {
-      component.phaseList = mockPhaseData;
-      const initialLength = component.phaseList.length;
-
-      component.addNewPhase();
-
-      expect(component.phaseList.length).toBe(initialLength + 1);
-      const newPhase = component.phaseList[component.phaseList.length - 1];
-      expect(newPhase.isNew).toBe(true);
-      expect(newPhase.editing).toBe(true);
-    });
-  });
-
-  describe('onlyPreviousPhase', () => {
-    it('should filter phases with year greater than or equal to current phase year', () => {
-      component.phaseList = [{ phase_year_ts: 2022 }, { phase_year_ts: 2023 }, { phase_year_ts: 2024 }];
-
-      const currentPhase = { phase_year_ts: 2023 };
-      const result = component.onlyPreviousPhase(currentPhase);
-
-      expect(result.length).toBe(2);
-      expect(result).toEqual([{ phase_year_ts: 2023 }, { phase_year_ts: 2024 }]);
-    });
-  });
-
-  describe('cancelAction', () => {
-    it('should cancel editing for existing phase', () => {
-      const phase = { id: 1, editing: true, isNew: false };
-
-      component.cancelAction(phase);
-
-      expect(phase.editing).toBe(false);
-    });
-
-    it('should remove new phase from the list when canceling', () => {
-      const newPhase = { id: undefined, editing: true, isNew: true };
-      component.phaseList = [mockPhaseData[0], newPhase];
-
-      component.cancelAction(newPhase);
-
-      expect(newPhase.editing).toBe(false);
-      expect(component.phaseList.length).toBe(1);
-      expect(component.phaseList[0]).toEqual(mockPhaseData[0]);
-    });
-  });
-
-  describe('disablePreviousYear', () => {
-    it('should return array of years from phase list', () => {
-      component.phaseList = [{ phase_year: 2023 }, { phase_year: 2024 }];
-
-      const result = component.disablePreviousYear();
-
-      // This method has a bug - it returns undefined due to forEach
-      // Testing the current behavior
-      expect(result).toBeUndefined();
-    });
-  });
-
-  describe('getPortfolios', () => {
-    it('should fetch and set portfolioList', () => {
-      component.getPortfolios();
-      expect(mockResultsApiService.GET_portfolioList).toHaveBeenCalled();
-      expect(component.portfolioList).toEqual(mockPortfolios);
-    });
-  });
-
-  describe('updateVariablesToSave', () => {
-    it('should copy portfolio_id to portfolio_id_ts', () => {
-      const phase: any = { portfolio_id: 2 };
-      component.updateVariablesToSave(phase);
-      expect(phase.portfolio_id_ts).toBe(2);
-    });
-  });
-
-  describe('updateMainVariables', () => {
-    it('should copy portfolio_id_ts to portfolio_id', () => {
-      const phase: any = { portfolio_id_ts: 3 };
-      component.updateMainVariables(phase);
-      expect(phase.portfolio_id).toBe(3);
-    });
-  });
-
-  describe('getMandatoryIncompleteFields', () => {
-    it('should return error message if portfolio_id_ts is missing', () => {
-      const phase = {
-        phase_name_ts: 'Test',
-        phase_year_ts: 2024,
-        toc_pahse_id_ts: 1,
-        start_date_ts: '2024-01-01',
-        end_date_ts: '2024-12-31',
-        portfolio_id_ts: null
-      };
-      const result = component.getMandatoryIncompleteFields(phase);
-      expect(result).toContain('<strong> Portfolio </strong> is required to create');
-    });
-    it('should return empty string if portfolio_id_ts is present', () => {
-      const phase = {
-        phase_name_ts: 'Test',
-        phase_year_ts: 2024,
-        toc_pahse_id_ts: 1,
-        start_date_ts: '2024-01-01',
-        end_date_ts: '2024-12-31',
-        portfolio_id_ts: 2
-      };
-      const result = component.getMandatoryIncompleteFields(phase);
-      expect(result).toBe('');
-    });
-  });
-});
